\name{donorAcceptorTagged}
\alias{donorAcceptorTagged}
\alias{cy005c}
\alias{cy006}
\docType{data}
\title{Fluorescent lifetime imaging microscopy (FLIM) data}
\description{Donor-and-acceptor tagged fluorescent lifetime
  imaging microscopy (FLIM) data.  }
\usage{
data("donorAcceptorTagged")
}
\format{
  \code{cy005c} and \code{cy006} are objects of class \code{dat}
  representing donor-and-acceptor tagged data.
}
\references{This data was described in

  Mullen KM, van Stokkum IHM (2008).
  The variable projection algorithm in time-resolved spectroscopy,
  microscopy and mass-spectroscopy applications,
  \emph{Numerical Algorithms}, \bold{in press}, 
\url{http://dx.doi.org/10.1007/s11075-008-9235-2}.
}
\details{
See \code{\link{FLIMplots}} for examples using this data. 
}

\keyword{datasets}
