\name{Internals}
\alias{Internals} 
\alias{addDscal} 
\alias{addPrel} 
\alias{addPrelCl}
\alias{applyWeighting} 
\alias{applyWeightingModel} 
\alias{baseCorlambda}
\alias{baseCortime} 
\alias{calcB} 
\alias{calcC} 
\alias{calcCirf}
\alias{calcEbspline} 
\alias{calcEhiergaus} 
\alias{compCoh} 
\alias{compModel}
\alias{decayirf} 
\alias{diffAdd} 
\alias{diffChange} 
\alias{diffFree}
\alias{diffRel} 
\alias{diffRemove} 
\alias{displayEst}
\alias{doClpConstr} 
\alias{doSVD}
\alias{fillK}
\alias{fullKF} 
\alias{gaus} 
\alias{getClpConstr}
\alias{getClpList} 
\alias{getCoh} 
\alias{getCohToPlot} 
\alias{getConToPlot}
\alias{getDiffTheta} 
\alias{getDiffThetaChange} 
\alias{getDiffThetaCl}
\alias{getDiffThetaClChange} 
\alias{getFixed}
\alias{getGroups} 
\alias{getModel} 
\alias{getPar} 
\alias{getPrel}
\alias{getPrelTheta} 
\alias{getSpecList}
\alias{getSpecToPlot} 
\alias{getTheta} 
\alias{getThetaCl} 
\alias{init.dat}
\alias{init.fc} 
\alias{init.het} 
\alias{init.kin} 
\alias{initModellist}
\alias{init.multimodel} 
\alias{init.multitheta} 
\alias{init.res}
\alias{init.spec} 
\alias{init.theta} 
\alias{irfparF} 
\alias{l2nu}
\alias{linloglines} 
\alias{linlogplot} 
\alias{matlinlogplot} 
\alias{normdat} 
\alias{plotFLIM} 
\alias{plotKinBreakDown}
\alias{plotKinSpec} 
\alias{plotSelectedSpectra} 
\alias{plotSelectedTraces}
\alias{plotSelectedTracesSuper} 
\alias{plotSpectra} 
\alias{plotTraces}
\alias{plotTracesSuper} 
\alias{rescomp} 
\alias{sample_sel} 
\alias{setGenerics}
\alias{set_getClpindepX} 
\alias{set_getClpindepX_kin}
\alias{set_getClpindepX_spec} 
\alias{setMethods} 
\alias{set_plotter_kin}
\alias{set_plotter_spec} 
\alias{set_residPart_kin}
\alias{set_residPart_spec} 
\alias{simndecay_gen} 
\alias{simpExp} 
\alias{simpPol} 
\alias{skew} 
\alias{specModel}
\alias{specparF} 
\alias{timpErrors} 
\alias{weightNL} 
\alias{weightPsi} 
\alias{weightSM} 
\alias{writeEst} 
\alias{writeFit} 
\alias{init.opt} 
\alias{init.kinopt}
\alias{init.specopt}
\alias{plotKinSpecEst}
\alias{plotEst}
\alias{init.fit}
\alias{fillResult}
\title{TIMP function used internally}
\description{TIMP function used internally}
\author{Katharine M. Mullen, Ivo H. M. van Stokkum}
\details{The functions linked to below are for direct use. }
\seealso{
\code{\link{baseIRF}},\code{\link{readData}}, 
\code{\link{preProcess}},\code{\link{initModel}},
\code{\link{fitModel}},\code{\link{examineFit}}, 
\code{\link{sumKinSpecEst}}  
}
\keyword{file}

