\name{mea_IRF}
\alias{mea_IRF}
\docType{data}
\title{Instrument response for fluorescent lifetime imaging microscopy (FLIM) data}
\description{A measured instrument response over 256 time channels for a
  set of fluorescent lifetime imaging microscopy (FLIM) datasets.}  
\usage{
data("mea_IRF")
}
\format{
  \code{mea_IRF} represents a measured instrument response as a numeric
  vector over 256 time channels. 
}
\references{This data was described in

  Mullen KM, van Stokkum IHM (2008).
  The variable projection algorithm in time-resolved spectroscopy,
  microscopy and mass-spectroscopy applications,
  \emph{Numerical Algorithms}, \bold{in press}, 
\url{http://dx.doi.org/10.1007/s11075-008-9235-2}.
}
\details{
See \code{\link{FLIMplots}} for examples using this data. 
}

\keyword{datasets}
