\name{target}
\alias{target}
\alias{C1}
\alias{C2}
\alias{C3}
\docType{data}
\title{Ultrafast time-resolved fluorescence data}
\description{Ultrafast time-resolved absorption data}  
\usage{
data("target")
}
\format{
  \code{C1}, \code{C2} and \code{C3} are objects of class \code{dat}.
}
\references{This data was described in
Bonetti C, Mathes T, van Stokkum IHM,
Mullen KM, Groot ML and 
van Grondelle R, Hegemann P and Kennis, JTM (2008),
The variable projection algorithm in time-resolved spectroscopy,
  microscopy and mass-spectroscopy applications,
  Hydrogen bond switching among flavin and amino acid side 
chains in the {BLUF} photoreceptor observed by ultrafast infrared spectroscopy,
  \emph{Biophysical Journal}, \bold{in press}, 
\url{http://www.biophysj.org/cgi/content/abstract/biophysj.108.139246v1}.
}
\details{
See \code{\link{preProcess}} for examples using this data. 
}

\keyword{datasets}
