\name{Internals}
\alias{Internals} 
\alias{addDscal} 
\alias{addPrel} 
\alias{addPrelCl}
\alias{applyWeighting} 
\alias{applyWeightingModel} 
\alias{baseCorlambda}
\alias{baseCortime} 
\alias{calcB} 
\alias{calcC} 
\alias{calcCirf}
\alias{calcEbspline} 
\alias{calcEhiergaus} 
\alias{compCoh} 
\alias{compModel}
\alias{compOsc}
\alias{convGausExp}
\alias{diffAdd} 
\alias{diffChange} 
\alias{diffFree}
\alias{diffRel} 
\alias{diffRemove} 
\alias{displayEst}
\alias{doClpConstr} 
\alias{doSVD}
\alias{fillK}
\alias{fullKF} 
\alias{gaus} 
\alias{getClpConstr}
\alias{getClpList} 
\alias{getCoh} 
\alias{getCohToPlot} 
\alias{getConToPlot}
\alias{getDiffTheta} 
\alias{getDiffThetaChange} 
\alias{getDiffThetaCl}
\alias{getDiffThetaClChange} 
\alias{getFixed}
\alias{getGroups} 
\alias{getModel} 
\alias{getOsc} 
\alias{getPar} 
\alias{getPrel}
\alias{getPrelTheta} 
\alias{getSpecList}
\alias{getSpecToPlot} 
\alias{getTheta} 
\alias{getThetaCl} 
\alias{initModellist}
\alias{irfparF} 
\alias{l2nu}
\alias{linloglines} 
\alias{linlogplot} 
\alias{matlinlogplot} 
\alias{normdat} 
\alias{plotFLIM} 
\alias{plotKinBreakDown}
\alias{plotClp} 
\alias{plotSelectedSpectra} 
\alias{plotSelectedTraces}
\alias{plotSelectedTracesSuper} 
\alias{plotTraces}
\alias{plotTracesSuper} 
\alias{rescomp} 
\alias{sample_sel} 
\alias{simndecay_gen} 
\alias{simpExp} 
\alias{simpPol} 
\alias{skew} 
\alias{specModel}
\alias{specparF} 
\alias{timpErrors} 
\alias{weightNL} 
\alias{weightPsi} 
\alias{weightSM} 
\alias{writeEst} 
\alias{writeFit} 
\alias{plotKinSpecEst}
\alias{plotEst}
\alias{fillResult}
\alias{processOrder}
\alias{multiLin}
\alias{scaleConList}
\alias{getXsuper}
\alias{getKinConcen}
\alias{doConstrSuper}
\alias{compModelMass}
\alias{barplot3}
\alias{getAutoClp0}
\alias{getResidRet}
\alias{linlogtics}
\alias{unitstep}
\alias{getSelectedTracesMax}
\alias{calcD} 
\title{TIMP function used internally}
\description{TIMP function used internally}
\author{Katharine M. Mullen and Joris J. Snellenburg and Ivo H. M. van Stokkum}
\seealso{
\code{\link{baseIRF}},\code{\link{readData}}, 
\code{\link{preProcess}},\code{\link{initModel}},
\code{\link{fitModel}},\code{\link{examineFit}}, 
\code{\link{sumKinSpecEst}}  
}
\keyword{internal}