\name{TIMP-package}
\docType{package}
\alias{TIMP-package}
\alias{TIMP}
\title{a problem solving environment for fitting separable nonlinear models in physics and chemistry applications}
\description{
 TIMP is a problem solving environment for fitting separable nonlinear models to measurements arising in physics and chemistry experiments, and has been extensively applied to time-resolved spectroscopy and FLIM-FRET data.
}
\author{Katharine M. Mullen, Ivo H. M. van Stokkum  
Maintainer: Katharine M. Mullen \email{kate@nat.vu.nl} 
}
\references{
  Mullen KM, van Stokkum IHM (2007). ``TIMP: An R Package for Modeling
  Multi-way Spectroscopic Measurements.'' Journal of Statistical
  Software, 18(3). URL \url{http://www.  jstatsoft.org/v18/i03/}.

  Laptenok S, Mullen KM, Borst JW, van Stokkum IHM, Apanasovich VV,
  Visser AJWG (2007). ``Fluorescence Lifetime Imaging Microscopy (FLIM)
  Data Analysis with TIMP.''  Journal of Statistical Software,
  18(8). URL \url{http://www.jstatsoft.org/v18/i08/}.
  
  See \url{ http://timp.r-forge.r-project.org/} or
  \url{http://www.nat.vu.nl/~kate/TIMP/} for further documentation.
}
\keyword{package}
